+++ src/clj/lambdacd/steps/control_flow.clj
@@ -4,9 +4,10 @@
             [clojure.core.async :as async]
             [lambdacd.steps.support :as support]
             [lambdacd.step-id :as step-id]
-            [lambdacd.steps.status :as status]
+
             [lambdacd.util.internal.temp :as temp-util]
-            [lambdacd.execution.internal.serial-step-result-producer :as serial-step-result-producer])
+            [lambdacd.execution.internal.serial-step-result-producer :as serial-step-result-producer]
+            [lambdacd.stepstatus.unify :as unify])
   (:refer-clojure :exclude [alias])
   (:import (java.util UUID)))
 
@@ -23,7 +24,7 @@
                                  (async/go-loop [statuses []]
                                    (if-let [result (async/<! all-resp-results-ch)]
                                      (let [new-statuses (conj statuses (:status result))]
-                                       (reset! unified-status (status/successful-when-one-successful new-statuses))
+                                       (reset! unified-status (unify/successful-when-one-successful new-statuses))
                                        (if (= :success (:status result))
                                          result
                                          (recur new-statuses))))))]
@@ -68,7 +69,7 @@
                                                       :is-killed kill-switch
                                                       :step-result-producer either-step-result-producer
                                                       :retrigger-predicate (constantly :rerun)
-                                                      :unify-results-fn (support/unify-only-status status/successful-when-one-successful))]
+                                                      :unify-results-fn (support/unify-only-status unify/successful-when-one-successful))]
       (reset! kill-switch true)
       (remove-watch parent-kill-switch watch-ref)
       (if (= :success (:status execute-output))
@@ -91,7 +92,7 @@
 (defn- execute-steps-in-parallel [steps args ctx]
   (execution/execute-steps steps args ctx
                            :step-result-producer parallel-step-result-producer
-                           :unify-results-fn (support/unify-only-status status/successful-when-all-successful)
+                           :unify-results-fn (support/unify-only-status unify/successful-when-all-successful)
                            :retrigger-predicate parallel-retrigger-predicate
                            :is-killed (:is-killed ctx)))
 
@@ -136,12 +137,12 @@
   (fn [args ctx]
     (post-process-container-results
       (execution/execute-steps steps (assoc args :cwd cwd) ctx
-                               :unify-results-fn (support/unify-only-status status/successful-when-all-successful)))))
+                               :unify-results-fn (support/unify-only-status unify/successful-when-all-successful)))))
 
 (defn- run-steps-in-sequence [args ctx steps]
   (post-process-container-results
     (execution/execute-steps steps args ctx
-                             :unify-results-fn (support/unify-only-status status/successful-when-all-successful)
+                             :unify-results-fn (support/unify-only-status unify/successful-when-all-successful)
                              :is-killed (:is-killed ctx))))